requirements.txt
langchain-community
fastapi
uvicorn
langchain
langchain-openai
langchain-pinecone
python-dotenv
